move pause logic to pause module
