+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Pause system moved to new pause module
 - Menu and pause state blinking moved to new effects module
 
 ## [0.0.7] - 2023-09-29
